+++ packages/studio-ui/src/web/views/Nlu/entities/PatternEntity.tsx
@@ -12,9 +12,9 @@ import {
 } from '@blueprintjs/core'
 import { NLU } from 'botpress/sdk'
 import { lang } from 'botpress/shared'
+import { getEntityId } from 'common/entity-id'
 import _ from 'lodash'
 import React, { useEffect, useState } from 'react'
-import { getEntityId } from './entity-id'
 
 import style from './style.scss'
